[Feature] Calendar checkin/checkout logic updated

+++ pages/book.tsx
@@ -3,10 +3,12 @@ import axios from 'axios';
 import Link from 'next/link';
 import { withRouter } from 'next/router';
 import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
+import { bindActionCreators, Dispatch } from 'redux';
+import * as dateFns from 'date-fns';
+
 import { User } from '../server/db/models/interfaces';
 import { bookTrip } from '../store/store';
-import { Dispatch } from 'redux';
+import Calendar from '../components/Calendar';
 
 interface Tripmates {
   [key: number]: any;
@@ -17,15 +19,18 @@ const Book = (props: any) => {
   const { listing } = trip;
 
   const initTripmates: Tripmates = {};
+  const formattedDateFrom: Date = new Date(trip.dateFrom);
+  const formattedDateTo: Date = new Date(trip.dateTo);
 
-  const [renderView, setRender] = useState(false);
+  // const [renderView, setRender] = useState(false);
   const [tripmates, setTripmates] = useState(initTripmates);
-  const [dateFrom, setDateFrom] = useState(trip.dateFrom);
-  const [dateTo, setDateTo] = useState(trip.dateTo);
+  const [dateFrom, setDateFrom] = useState(formattedDateFrom);
+  const [dateTo, setDateTo] = useState(formattedDateTo);
+  const [bookError, setBookError] = useState('');
 
   useEffect(() => {
     if (!listing) router.replace('/listings');
-    setRender(true);
+    // setRender(true);
   }, []);
 
   const handleSelect = (e: React.FormEvent, user: User) => {
@@ -41,25 +46,31 @@ const Book = (props: any) => {
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
 
-    const userIds: number[] = [];
-    const newTrip = {
-      dateFrom,
-      dateTo,
-      status: 'booked',
-      listingId: listing.id,
-    };
+    if (dateFrom && dateFrom) {
+      const userIds: number[] = [];
+      const newTrip = {
+        dateFrom: dateFns.format(dateFrom, 'yyyy-MM-dd'),
+        dateTo: dateFns.format(dateTo, 'yyyy-MM-dd'),
+        status: 'booked',
+        listingId: listing.id,
+      };
 
-    Object.keys(tripmates).forEach((idStr: string) => {
-      const id = parseInt(idStr);
-      if (tripmates[id]) userIds.push(id);
-    });
+      Object.keys(tripmates).forEach((idStr: string) => {
+        const id = parseInt(idStr);
+        if (tripmates[id]) userIds.push(id);
+      });
 
-    const res = await axios.post(`/api/trips`, { userIds, trip: newTrip });
-    const resTrip = res.data;
+      const res = await axios.post(`/api/trips`, { userIds, trip: newTrip });
+      const resTrip = res.data;
 
-    bookTrip(resTrip);
-    await axios.delete(`api/trips/${trip.id}`);
-    router.push('/confirmation');
+      bookTrip(resTrip);
+      await axios.delete(`api/trips/${trip.id}`);
+      router.push('/confirmation');
+    } else {
+      setBookError(
+        "Please make sure you've properly selected a checkin and checkout date"
+      );
+    }
   };
 
   return (
@@ -73,15 +84,15 @@ const Book = (props: any) => {
         <p>{listing.description}</p>
       </div>
 
-      <form name='tripmate-selection' onSubmit={handleSubmit}>
+      <form name="tripmate-selection" onSubmit={handleSubmit}>
         <h4>Choose your tripmates!</h4>
         <ul>
           {users.map((user: any) => {
             const fullName = `${user.firstName} ${user.lastName}`;
             return (
-              <div key={user.id} className='tripmate-option'>
+              <div key={user.id} className="tripmate-option">
                 <input
-                  type='checkbox'
+                  type="checkbox"
                   name={user.id}
                   onChange={(e) => handleSelect(e, user)}
                 />
@@ -91,12 +102,23 @@ const Book = (props: any) => {
                 </label>
               </div>
             );
-            // <li key={user.id}>{`${user.firstName} ${user.lastName}`}</li>;
           })}
         </ul>
 
         <h4>Confirm the dates</h4>
-        <input
+        <Calendar
+          checkin={dateFrom}
+          setCheckin={(v: any) => {
+            setDateFrom(v);
+            setBookError('');
+          }}
+          checkout={dateTo}
+          setCheckout={(v: any) => {
+            setDateTo(v);
+            setBookError('');
+          }}
+        />
+        {/* <input
           type='date'
           name='date-from'
           defaultValue={dateFrom}
@@ -107,10 +129,10 @@ const Book = (props: any) => {
           name='date-to'
           defaultValue={dateTo}
           onChange={(e) => setDateTo(e.target.value)}
-        />
-        <button type='submit'>Confirm Booking</button>
+        /> */}
+        <button type="submit">Confirm Booking</button>
       </form>
-      <Link href='/listings'>
+      <Link href="/listings">
         <button>Back to listings</button>
       </Link>
     </div>
